Fix crawlable headings on money pages
