+++ src/app/[locale]/book-assessment/layout.tsx
@@ -72,4 +72,3 @@ export default async function BookAssessmentLayout({
     </>
   )
 }
-
